renaming err to debug package
